# server/models/users.ts
1
import { Model, DataTypes } from 'sequelize';
import sequelize from '../dataAccess/dataAccess';
  

class User extends Model {
  public id!: number;
  public userName!: string;
  public userPassword!: string;
  public userEmail!:string;
  public userToken!: string;
}

  User.init( {
    id: {
        primaryKey: true,
        type: DataTypes.INTEGER,
        autoIncrement: true,
        allowNull: false
    },
    userName: {
        type: DataTypes.STRING,
        allowNull: false
    },
    userPassword: {
        type: DataTypes.STRING,
        allowNull: false
    },
    userEmail: {
      type: DataTypes.STRING,
      allowNull: false,
      unique: true
    },
    userToken: {
        type: DataTypes.STRING,
        allowNull: false
      }
    
} ,{
    sequelize,
    tableName: 'users',
});
export {User}